third_party/blink/web_tests/fast/css/fontface-methods-expected.txt

CONSOLE WARNING: Failed to decode downloaded font: data:font/truetype;base64,
This is a testharness.js-based test.
Harness: the test ran to completion.
